Cambios menores en los controladores

+++ BACKEND/app/Http/Controllers/RolsController.php
@@ -4,26 +4,29 @@ namespace App\Http\Controllers;
 
 use App\Models\Rols;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use PhpParser\Node\Expr\BinaryOp\Concat;
 
 class RolsController extends Controller
 {
 
-    public function index()
+    public function index(Request $request)
     {
         return response()->json([Rols::all()], 200);
     }
 
     public function store(Request $request)
     {
-        $rol = new Rols();
-        $rol->name = $request->input('name');
-        $rol->description = $request->input('description');
-        $rol->save();
-        return response()->json([
-            'message' => 'Rol creado exitosamente',
-            'res' => true
-        ], 201);
+        if (Auth::user() &&  Auth::user()->rol_id == 1) {
+            $rol = new Rols();
+            $rol->name = $request->input('name');
+            $rol->description = $request->input('description');
+            $rol->save();
+            return response()->json([
+                'message' => 'Rol creado exitosamente',
+                'res' => true
+            ], 201);
+        }
     }
 
     public function show($id)
@@ -37,30 +40,34 @@ class RolsController extends Controller
 
     public function update(Request $request)
     {
-        $id = $request->input('id');
-        $rol = Rols::find($id);
-        if (!$rol) {
-            return response()->json(['Rol no existente'], 404);
+        if (Auth::user() &&  Auth::user()->rol_id == 1) {
+            $id = $request->input('id');
+            $rol = Rols::find($id);
+            if (!$rol) {
+                return response()->json(['Rol no existente'], 404);
+            }
+            $rol->name = $request->input('name');
+            $rol->description = $request->input('description');
+            $rol->save();
+            return response()->json([
+                'message' => 'Rol actualizado exitosamente',
+                'res' => true
+            ], 200);
         }
-        $rol->name = $request->input('name');
-        $rol->description = $request->input('description');
-        $rol->save();
-        return response()->json([
-            'message' => 'Rol actualizado exitosamente',
-            'res' => true
-        ], 200);
     }
 
     public function destroy($id)
     {
-        $rol = Rols::find($id);
-        if (!$rol) {
-            return response()->json(['Rol no existente'], 404);
+        if (Auth::user() &&  Auth::user()->rol_id == 1) {
+            $rol = Rols::find($id);
+            if (!$rol) {
+                return response()->json(['Rol no existente'], 404);
+            }
+            $rol->delete();
+            return response()->json([
+                'message' => 'Rol eliminado',
+                $rol
+            ], 200);
         }
-        $rol->delete();
-        return response()->json([
-            'message' => 'Rol eliminado',
-            $rol
-        ], 200);
     }
 }
